+++ app/keygen/internal/decrypt_test.go
@@ -8,4 +8,4 @@
 >/'  SPDX-License-Identifier: BSD-2-Clause
 */
 
-package main
+package internal
